Organize record model around Earth4D world mesh
