+++ core/__init__.py
@@ -1,4 +1,1 @@
-"""Production world model, data adapter, and fixed-step trainer.
-
-Import :mod:`deepearth.core.fusion` directly to avoid eagerly loading the CUDA hash encoder.
-"""
+"""DeepEarth world state, fusion, data, and training."""
